+++ scratch.js
@@ -24,19 +24,12 @@ let txt = ''
 // console.log(doc.text(opts))
 
 
+let doc = nlp("before (match), after")
+let out = doc.wrap({
+  'match': () => `few more words`,
+})
+console.log(out)
 
-
-// const doc = nlp("foo. found one after found bad")
-// let net = nlp.buildNet([{ match: 'found .', ifNo: ['bad'] }])
-// doc.match(net).debug()
-
-nlp(`were based out of`).debug()
-
-
-// let doc = nlp('the remaining claims fail').debug()
-// let doc = nlp('The service is fast psych')
-// let net = nlp.buildNet([{ match: 'is fast .', notIf: 'psych' }])
-// doc.match(net).debug()
 
 let arr = [