+++ src/3-three/nouns/api/toPlural.js
@@ -2,7 +2,7 @@ const keep = { tags: true }
 
 const hasPlural = function (parsed) {
   let { root } = parsed
-  if (root.has('^(#Uncountable|#Possessive|#ProperNoun|#Place|#Pronoun)+$')) {
+  if (root.has('^(#Uncountable|#Possessive|#ProperNoun|#Place|#Pronoun|#Acronym)+$')) {
     return false
   }
   return true
@@ -26,7 +26,17 @@ const nounToPlural = function (m, parsed) {
   // should we change the determiner/article?
   if (parsed.determiner.has('(a|an)')) {
     // 'a captain' -> 'the captains'
-    m.replace(parsed.determiner, 'the', keep)
+    // m.replace(parsed.determiner, 'the', keep)
+    m.remove(parsed.determiner)
+  }
+  // should we change the following copula?
+  let copula = parsed.root.after('not? #Adverb+? [#Copula]', 0)
+  if (copula.found) {
+    if (copula.has('is')) {
+      m.replace(copula, 'are')
+    } else if (copula.has('was')) {
+      m.replace(copula, 'were')
+    }
   }
   return m
 }