+++ changelog.md
@@ -6,9 +6,14 @@ compromise uses semver, and pushes to npm and github frequently
 
 While all _Major_ releases should be reviewed, our only _large_ releases are **v6** in 2016 **v12** in 2019 and **v14** in 2022. Others have been mostly incremental.
 
+<!-- #### [planned breaking]
+- fix doc.json(0) inconsistency
+-->
+
 <!-- #### 14.6.0 [Unreleased]
 - **[new]** - match term id
 - **[change]** - tag text by default on .concat('')
+- **[change]** - allow changing term prePunctuation
 - **[new]** - .wrap() method
 -->